+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 2025-11-30
+
+- 不再使用 nvim
+- 不再使用 micro (ctrl q 快捷键在 Cursor/VS Code 终端中不好使)，编辑器现在只 vim
+
 ## 2025-11-17
 
 - zsh 编辑命令使用 vim (因为 micro 在 vscode/Cursor 终端中 ctrl q 可能已经被占用，而且 micrio 编辑命令不方便)